handle returning to form hierarchy activity after process is killed in background
